tinygo: enable decoding CBOR tag to Go interface
This commit enables decoding CBOR tag to Go interface when
compiling with TinyGo, because TinyGo v0.41.0 implemented
a required feature.

Previously, when compiling with TinyGo, we disabled the ability
to decode a registered CBOR tag into a Go interface value because
older versions of TinyGo didn't implement reflect.Type.AssignableTo
(used internally by reflect.Type.Implements).

More details at https://github.com/tinygo-org/tinygo/issues/4277.

TinyGo v0.41.0 added support for reflect.Type.AssignableTo
in https://github.com/tinygo-org/tinygo/pull/5304.
